# Conflicts:
#	APP0200025/APP0200025.csproj

+++ APP0200025/Controllers/ChoTiepNhanKetQuaController.cs
@@ -50,7 +50,7 @@ namespace APP0200025.Controllers
             clHangHoa.UpdateNguoiXem(iID_MaHangHoa, User.Identity.Name);
 
             ViewData["menu"] = 235;
-            HangHoaModels models = clHangHoa.GetHangHoaById(Convert.ToInt32(iID_MaHangHoa));
+            HangHoaModels models = clHangHoa.GetHangHoaById(Convert.ToInt64(iID_MaHangHoa));
 
             return View(models);
         }
